+++ R/options.R
@@ -1,10 +1,9 @@
 #' Load roxygen2 options
 #'
 #' @description
-#' Options can be stored in the `Roxygen` field of the `DESCRIPTION`, or
-#' in `man/roxygen/meta.R`. In either case, the code is parsed and evaluated
-#' in a child of the base environment. Call `roxy_meta_get()` to access
-#' current option values from within tag and roclet methods.
+#' Options can be stored in `DESCRIPTION` using `Config/roxygen2/` fields,
+#' or in `man/roxygen/meta.R`. Call `roxy_meta_get()` to access current
+#' option values from within tag and roclet methods.
 #'
 #' Options in `man/roxygen/meta.R` override those present in `DESCRIPTION`.
 #'
@@ -35,18 +34,20 @@
 #'   manual. (This only applies to images supplied via markdown.)
 #'
 #' @section How to set:
-#' Either set in `DESCRIPTION`:
+#' Either set in `DESCRIPTION` using `Config/roxygen2/` fields:
 #'
 #' ```
-#' Roxygen: list(markdown = TRUE, load = "installed")
+#' Config/roxygen2/markdown: TRUE
+#' Config/roxygen2/load: installed
 #' ```
 #'
-#' Or if longer, you can put in `/man/roxygen/meta.R`:
+#' Or if you need more complex options (like `rd_family_title` or
+#' `knitr_chunk_options`), put them in `man/roxygen/meta.R`:
 #'
 #' ```
 #' list(
-#'   markdown = TRUE,
-#'   load = "installed"
+#'   rd_family_title = list(models = "Model functions"),
+#'   knitr_chunk_options = list(fig.width = 7)
 #' )
 #' ```
 #'
@@ -54,7 +55,9 @@
 #' @export
 #' @family extending
 load_options <- function(base_path = ".") {
-  desc <- load_options_description(base_path)
+  old <- load_options_roxygen(base_path)
+  config <- load_options_config(base_path)
+  desc <- utils::modifyList(old, config)
   meta <- load_options_meta(base_path)
   opts <- utils::modifyList(desc, meta)
 
@@ -87,7 +90,7 @@ load_options <- function(base_path = ".") {
   utils::modifyList(defaults, opts)
 }
 
-load_options_description <- function(base_path = ".") {
+load_options_roxygen <- function(base_path = ".") {
   desc_path <- file.path(base_path, "DESCRIPTION")
   dcf <- read.dcf(desc_path, fields = c("Roxygen", "Package", "LazyData"))
   desc_opts <- dcf[[1, 1]]
@@ -102,6 +105,36 @@ load_options_description <- function(base_path = ".") {
   opts$current_package_dir <- normalizePath(base_path)
   opts$lazy_data <- identical(dcf[[1, 3]], "true")
   opts
+}
+
+config_fields <- c(
+  "markdown",
+  "load",
+  "old_usage",
+  "r6",
+  "restrict_image_formats",
+  "packages",
+  "roclets"
+)
+
+load_options_config <- function(base_path = ".") {
+  fields <- paste0("Config/roxygen2/", config_fields)
+  values <- desc::desc_get(fields, file = base_path)
+  names(values) <- config_fields
+
+  values <- values[!is.na(values)]
+  lapply(values, parse_config_value)
+}
+
+parse_config_value <- function(x) {
+  values <- scan(
+    text = x,
+    what = "character",
+    sep = ",",
+    strip.white = TRUE,
+    quiet = TRUE
+  )
+  utils::type.convert(values, as.is = TRUE)
 }
 
 load_options_meta <- function(base_path = ".", path = "man/roxygen/meta.R") {