+++ tests/testthat/test-rd-r6-methods-self.R
@@ -52,6 +52,23 @@ test_that("warns about multiple @return(s) tags", {
   expect_equal(docs$methods$self[[1]]$return, "First.")
 })
 
+test_that("class-level @param x,y inherited by methods (#1600)", {
+  text <- "
+    #' Class
+    #' @param x,y Numbers.
+    C <- R6::R6Class('C', cloneable = FALSE,
+      public = list(
+        #' @description Method.
+        meth = function(x, y) {}
+      )
+    )"
+  expect_silent(docs <- r6_doc(text))
+  expect_equal(
+    docs$methods$self[[1]]$params,
+    list(list(name = "x, y", description = "Numbers."))
+  )
+})
+
 test_that("warns about undocumented params", {
   text <- "
     #' Class